+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## [68.0.4](https://github.com/elastic/elastic-charts/compare/v68.0.3...v68.0.4) (2024-12-11)
+
+
+### Bug Fixes
+
+* **xy:** compute per series and global minPointsDistance ([#2571](https://github.com/elastic/elastic-charts/issues/2571)) ([8dae2c1](https://github.com/elastic/elastic-charts/commit/8dae2c1f4c99146aa757b2d3eec9d72846248cc7))
+
+
+### Performance Improvements
+
+* fix unnecessary re-render ([#2573](https://github.com/elastic/elastic-charts/issues/2573)) ([feacfd6](https://github.com/elastic/elastic-charts/commit/feacfd6247b9580a8d32bc5d6284329b2035c1ba))
+
+## [68.0.3](https://github.com/elastic/elastic-charts/compare/v68.0.2...v68.0.3) (2024-11-28)
+
+
+### Bug Fixes
+
+* **xy:** line annotation marker aria label fix ([#2558](https://github.com/elastic/elastic-charts/issues/2558)) ([3b056c3](https://github.com/elastic/elastic-charts/commit/3b056c395c5ff87044ea989a3b737022cea20509))
+
 ## [68.0.2](https://github.com/elastic/elastic-charts/compare/v68.0.1...v68.0.2) (2024-10-24)